06-service-recommender/backend/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   httpx
    #   openai
arize-phoenix-otel==0.10.3
    # via backend (pyproject.toml)
attrs==25.3.0
    # via
    #   jsonschema
    #   referencing
backoff==2.2.1
    # via posthog
certifi==2025.4.26
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.2
    # via requests
distro==1.9.0
    # via
    #   openai
    #   posthog
filetype==1.2.0
    # via haystack-experimental
googleapis-common-protos==1.70.0
    # via
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
grpcio==1.73.0
    # via opentelemetry-exporter-otlp-proto-grpc
h11==0.16.0
    # via httpcore
haystack-ai==2.14.2
    # via
    #   backend (pyproject.toml)
    #   haystack-experimental
haystack-experimental==0.10.0
    # via haystack-ai
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via openai
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
importlib-metadata==8.7.0
    # via opentelemetry-api
jinja2==3.1.6
    # via haystack-ai
jiter==0.10.0
    # via openai
jsonschema==4.24.0
    # via haystack-ai
jsonschema-specifications==2025.4.1
    # via jsonschema
lazy-imports==1.0.0
    # via haystack-ai
markupsafe==3.0.2
    # via jinja2
more-itertools==10.7.0
    # via haystack-ai
networkx==3.5
    # via haystack-ai
numpy==2.3.0
    # via haystack-ai
openai==1.85.0
    # via
    #   backend (pyproject.toml)
    #   haystack-ai
openinference-instrumentation==0.1.32
    # via
    #   arize-phoenix-otel
    #   openinference-instrumentation-haystack
openinference-instrumentation-haystack==0.1.24
    # via backend (pyproject.toml)
openinference-semantic-conventions==0.1.17
    # via
    #   arize-phoenix-otel
    #   openinference-instrumentation
    #   openinference-instrumentation-haystack
opentelemetry-api==1.34.0
    # via
    #   openinference-instrumentation
    #   openinference-instrumentation-haystack
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
    #   opentelemetry-instrumentation
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
opentelemetry-exporter-otlp==1.34.0
    # via arize-phoenix-otel
opentelemetry-exporter-otlp-proto-common==1.34.0
    # via
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
opentelemetry-exporter-otlp-proto-grpc==1.34.0
    # via opentelemetry-exporter-otlp
opentelemetry-exporter-otlp-proto-http==1.34.0
    # via opentelemetry-exporter-otlp
opentelemetry-instrumentation==0.55b0
    # via openinference-instrumentation-haystack
opentelemetry-proto==1.34.0
    # via
    #   arize-phoenix-otel
    #   opentelemetry-exporter-otlp-proto-common
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
opentelemetry-sdk==1.34.0
    # via
    #   arize-phoenix-otel
    #   openinference-instrumentation
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
opentelemetry-semantic-conventions==0.55b0
    # via
    #   arize-phoenix-otel
    #   openinference-instrumentation-haystack
    #   opentelemetry-instrumentation
    #   opentelemetry-sdk
packaging==25.0
    # via opentelemetry-instrumentation
posthog==4.6.0
    # via haystack-ai
protobuf==5.29.5
    # via
    #   googleapis-common-protos
    #   opentelemetry-proto
pydantic==2.11.5
    # via
    #   haystack-ai
    #   openai
pydantic-core==2.33.2
    # via pydantic
python-dateutil==2.9.0.post0
    # via
    #   haystack-ai
    #   posthog
python-dotenv==1.1.0
    # via backend (pyproject.toml)
pyyaml==6.0.2
    # via haystack-ai
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.4
    # via
    #   haystack-ai
    #   opentelemetry-exporter-otlp-proto-http
    #   posthog
rpds-py==0.25.1
    # via
    #   jsonschema
    #   referencing
six==1.17.0
    # via
    #   posthog
    #   python-dateutil
sniffio==1.3.1
    # via
    #   anyio
    #   openai
tenacity==9.1.2
    # via haystack-ai
tqdm==4.67.1
    # via
    #   haystack-ai
    #   openai
typing-extensions==4.14.0
    # via
    #   anyio
    #   arize-phoenix-otel
    #   haystack-ai
    #   openai
    #   openinference-instrumentation-haystack
    #   opentelemetry-api
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
    #   pydantic
    #   pydantic-core
    #   referencing
    #   typing-inspection
typing-inspection==0.4.1
    # via pydantic
urllib3==2.4.0
    # via requests
wrapt==1.17.2
    # via
    #   arize-phoenix-otel
    #   openinference-instrumentation-haystack
    #   opentelemetry-instrumentation
zipp==3.23.0
    # via importlib-metadata
